requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
